Move integration tests, too
